react-launchpad/Backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-generativeai